requirements.txt
pytesseract==0.3.10
Pillow>=10.4.0
pdf2image==1.16.3
tkinterdnd2>=0.3.0
opencv-python>=4.12.0
PyPDF2
pymupdf